backend/B3Functions/requirements.txt
azure-functions
azure-storage-blob==12.26.0
python-dotenv==1.1.1
pyodbc
requests==2.32.5
lxml==6.0.2
azure-core==1.35.0
azure-identity==1.24.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
dotenv==0.9.9
idna==3.10
isodate==0.7.2
msal==1.33.0
msal-extensions==1.3.1
pycparser==2.23
PyJWT==2.10.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0